bugfixes and documentation
changed the oak hideshow data in the champion room because apparently in a fresh run it works fine
repruposed the item that I attempted to deprecate in victory road
psyduck get access now to amnesia.
minor edits to the lvup moves. Aerodactyl now goes for the throat
Added the type matchup table as documentation